+++ playwright/tests/keywordSearch.spec.ts
@@ -4,7 +4,6 @@ import { CustomWorld } from "../support/setup";
 
 When("I input a search term", async function (this: CustomWorld) {
   await this.page.locator('[aria-label="Item Search"]').fill("IBM 1401");
-  await this.page.locator("#searchbar-button-search-bar").click();
 });
 
 Then(